requirements.txt
fastapi[standard]
python-dotenv
groq
supabase
psycopg2-binary